Change color argument types from int to double to simplify API usage

+++ Source/Exceptions/ColorModeValueRangeException.cs
@@ -4,9 +4,9 @@ using DrawingBoardNET.Drawing.Constants;
 
 namespace DrawingBoardNET.Exceptions;
 
-public class ColorModeValueRangeException(string component, int value, int maxValue, DBColorMode colorMode) : ArgumentOutOfRangeException(WriteErrorMessage(component, value, maxValue, colorMode))
+public class ColorModeValueRangeException(string component, double value, double maxValue, DBColorMode colorMode) : ArgumentOutOfRangeException(WriteErrorMessage(component, value, maxValue, colorMode))
 {
-    private static string WriteErrorMessage(string component, int value, int maxValue, DBColorMode colorMode) =>
+    private static string WriteErrorMessage(string component, double value, double maxValue, DBColorMode colorMode) =>
         $"\nError, the value of {component} should be between 0 and {maxValue} " +
         $"in {colorMode} color mode\n" +
         $"\t{component} = {value}\n";
